feat: create and implement LayoutListComponent
